Update plugin and test with java 21 (#336)
* Modernize plugin

* Upgrade HtmlUnit from 2.x to 3.x

* Use most recent bom-2.387.x release

* Use lombok compatible with Java 21

* Remove trailing space

* Test with Java 21

Java 21 was released Sep 19, 2023. We want to announce full support
for Java 21 in early October and would like the most used plugins to be
compiled and tested with Java 21.

The acceptance test harness and plugin bill of materials tests are already
passing with Java 21. This is a further step to improve plugin readiness
for use with Java 21 and for development with Java 21.

The change intentionally tests only two Java configurations, Java 17
and Java 21 because we believe that the risk of a regression that only
affects Java 11 is low. We generate Java 11 byte code with the Java 17
and the Java 21 builds, so we're already testing Java 11 byte code.

---------

Co-authored-by: Basil Crow <me@basilcrow.com>
Co-authored-by: Tim Jacomb <timjacomb1@gmail.com>